first attempt
Added a cardFaces array with two images, and created card randomizing functions, to try to generate random card faces for the tiles.

So far, the .tile on click function works to hide the geometric background and display a random card face. The problem is that the card faces are sometimes duplicated.