thirdCardClicked
The countMoves() function is called in the selectTile click event, so the mismatchedPairs variable was increased by one each time a (third, fourth, fifth, etc.) tile was clicked within the 2.5s that mismatched pairs remain in the selectedPair array. To maintain an accurate number of totalMoves, I modified the countMoves() if statement to stipulate that mismatchedPairs can only increase if thirdCardClicked == false.